+++ src/components/chatroom/ChatRoomMain.tsx
@@ -1,9 +1,7 @@
-import { useState, useRef, useEffect } from 'react';
-import styled, { css } from 'styled-components';
+import { useRef, useEffect } from 'react';
+import styled from 'styled-components';
 import { Chat, Chats } from '../../types/index';
-import me from '../../data/me.json';
-import friends from '../../data/friends.json';
-import { getTimeStamp } from '../../utils/getTimeStamp';
+import ChatItem from './ChatItem';
 
 type ChatRoomMainProps = {
   partnerUserId: number;
@@ -11,10 +9,6 @@ type ChatRoomMainProps = {
 };
 
 function ChatRoomMain({ partnerUserId, chatList }: ChatRoomMainProps) {
-  const partner = friends.filter(
-    (friend) => friend.userId === partnerUserId
-  )[0];
-
   const scrollRef = useRef<HTMLDivElement>(null);
   useEffect(() => {
     scrollRef.current!.scrollTo({
@@ -26,18 +20,13 @@ function ChatRoomMain({ partnerUserId, chatList }: ChatRoomMainProps) {
   return (
     <ChatRoomMainBlock ref={scrollRef}>
       {chatList.map(({ userId, msg, unixTime }: Chat, idx: number) => (
-        <ChatItem key={idx} userId={userId}>
-          <img src={`${process.env.PUBLIC_URL}/imgs/${userId}.jpg`} />
-          <TextWrapper>
-            <UserName userId={userId}>
-              {userId === me.userId ? me.userName : partner.userName}
-            </UserName>
-            <MsgWrapper userId={userId}>
-              <Bubble userId={userId}>{msg}</Bubble>
-              <Timestamp>{getTimeStamp(unixTime)}</Timestamp>
-            </MsgWrapper>
-          </TextWrapper>
-        </ChatItem>
+        <ChatItem
+          key={idx}
+          partnerUserId={partnerUserId}
+          userId={userId}
+          msg={msg}
+          unixTime={unixTime}
+        />
       ))}
     </ChatRoomMainBlock>
   );
@@ -60,85 +49,6 @@ const ChatRoomMainBlock = styled.div`
     border-radius: 20px;
     border: 8px solid transparent;
   }
-`;
-
-const ChatItem = styled.div<{ userId: number }>`
-  margin: 3%;
-
-  display: flex;
-  flex-direction: ${({ userId }) =>
-    userId === me.userId ? 'row-reverse' : 'row'};
-
-  img {
-    width: 40px;
-    height: 40px;
-    margin: 1%;
-
-    border-radius: 50%;
-    object-fit: cover;
-  }
-`;
-
-const TextWrapper = styled.div`
-  width: 83%;
-  margin-left: 1%;
-  margin-right: 1%;
-
-  display: flex;
-  flex-direction: column;
-`;
-
-const UserName = styled.span<{ userId: number }>`
-  font-size: 0.7rem;
-  color: #3f464db2;
-
-  display: flex;
-  flex-direction: ${({ userId }) =>
-    userId === me.userId ? 'row-reverse' : 'row'};
-`;
-
-const MsgWrapper = styled.div<{ userId: number }>`
-  width: 100%;
-
-  display: flex;
-  flex-direction: ${({ userId }) =>
-    userId === me.userId ? 'row-reverse' : 'row'};
-`;
-
-const Bubble = styled.div<{ userId: number }>`
-  padding: 3.5%;
-  margin-top: 2%;
-
-  font-size: 0.7rem;
-  line-height: 1.1rem;
-
-  display: flex;
-  ${({ userId }) =>
-    userId === me.userId
-      ? css`
-          flex-direction: row-reverse;
-          color: #ffffffe9;
-          background: #1986fc;
-          border-radius: 12px 1px 12px 12px;
-        `
-      : css`
-          flex-direction: 'row';
-          color: #000000e5;
-          background: #f1f1f3;
-          border-radius: 1px 12px 12px 12px;
-        `}
-`;
-
-const Timestamp = styled.span`
-  color: #c8c8c8;
-  font-size: 0.5rem;
-
-  margin-left: 1%;
-  margin-right: 1%;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
 `;
 
 export default ChatRoomMain;
